+++ perf/cmake/windows.cmake
@@ -75,6 +75,12 @@ add_custom_command(
   TARGET ${PERF_APP}
   POST_BUILD
   COMMAND
+    ${CMAKE_COMMAND} -E copy "${CMAKE_CURRENT_SOURCE_DIR}/imb-stat-avg.pl"
+    $<TARGET_FILE_DIR:${PERF_APP}>)
+add_custom_command(
+  TARGET ${PERF_APP}
+  POST_BUILD
+  COMMAND
     ${CMAKE_COMMAND} -E copy "${CMAKE_CURRENT_SOURCE_DIR}/imb-speed.py"
     $<TARGET_FILE_DIR:${PERF_APP}>)
 add_custom_command(